Switch list item to take an argument for details
