+++ src/web/views/EmployerList/EmployerList.tsx
@@ -7,7 +7,7 @@ import { LocalizedStrings } from "../../../common/LocalizedStrings";
 import { getEmployersList } from "../../state/ducks/employers/selectors";
 import { getStrings } from "../../state/ducks/localization/selectors";
 
-import EmployerListItemDetailed from "../EmployerListItemDetailed/EmployerListItemDetailed";
+import EmployerListItem from "../EmployerListItem/EmployerListItem";
 import { EmployerListSearchFilter } from "../EmployerListSearch/EmployerListSearchFilter";
 
 import "./EmployerList.scss";
@@ -54,7 +54,7 @@ const EmployerList: React.FC<Props> = (props: Props): React.ReactElement => {
 		<div className="EmployerList__Container">
 			{filteredEmployers.map((e: EmployerRecordMetadata, i: number): JSX.Element => (
 				<div className="EmployerList__Item" key={`${i}-${e.id}`}>
-					<EmployerListItemDetailed employerId={e.id} />
+					<EmployerListItem employerId={e.id} showDetails={true} />
 				</div>
 			))}
 		</div>
