+++ src/web/views/EmployerListItem/EmployerListItem.tsx
@@ -1,7 +1,9 @@
 import React from "react";
+import ReactMarkdown from "react-markdown";
 import { useSelector } from "react-redux";
 import { RouteProps, useHistory } from "react-router-dom";
 
+import { DesignHelpers } from "../../../common/DesignHelpers";
 import { EmployerRecordMetadata } from "../../../common/EmployerRecordMetadata";
 import { LocalizedStrings } from "../../../common/LocalizedStrings";
 
@@ -9,6 +11,8 @@ import { AppState } from "../../state/AppState";
 import { getEmployerMetadata } from "../../state/ducks/employers/selectors";
 import { getStrings } from "../../state/ducks/localization/selectors";
 
+import EmployerActionLinks from "../EmployerActionLinks/EmployerActionLinks";
+import EmployerDetailsHeader from "../EmployerDetailsHeader/EmployerDetailsHeader";
 import EmployerLogo from "../EmployerLogo/EmployerLogo";
 
 import "./EmployerListItem.scss";
@@ -17,6 +21,8 @@ interface Props extends RouteProps {
 	employerId: string;
 
 	onClick?: () => void;
+
+	showDetails: boolean;
 }
 
 const EmployerListItem: React.FC<Props> = (props: Props): React.ReactElement | null => {
@@ -24,7 +30,7 @@ const EmployerListItem: React.FC<Props> = (props: Props): React.ReactElement | n
 
 	const { push } = useHistory();
 
-	const { employerId, onClick } = props;
+	const { employerId, onClick, showDetails } = props;
 
 	const employer: EmployerRecordMetadata | undefined =
 		useSelector((state: AppState) => getEmployerMetadata(state, employerId));
@@ -35,13 +41,32 @@ const EmployerListItem: React.FC<Props> = (props: Props): React.ReactElement | n
 
 	const onClickEvent: () => void = onClick || ((): void => push(`/employers/${employer.id}`));
 
+	if (!showDetails) {
+		return (
+			<div
+				className={`EmployerListItem__Container EmployerListItem__Rating--${employer.rating}`}
+				onClick={onClickEvent}
+			>
+				<EmployerLogo employer={employer} />
+				{employer.name}
+			</div>
+		);
+	}
+
 	return (
-		<div
-			className={`EmployerListItem__Container EmployerListItem__Rating--${employer.rating}`}
-			onClick={onClickEvent}
-		>
-			<EmployerLogo employer={employer} />
-			{employer.name}
+		<div className="EmployerListItem__Container">
+			<EmployerDetailsHeader employer={employer} onClickEmployerName={onClickEvent} useShortText={true} />
+			<div className="EmployerListItem__Summary" onClick={onClickEvent}>
+				<ReactMarkdown source={employer.summary} />
+				<div className="EmployerListItem__OverflowScreen" />
+			</div>
+			<div className="EmployerListItem__Actions">
+				<EmployerActionLinks employer={employer} />
+				<a className="EmployerListItem__ReadMore" onClick={onClickEvent}>
+					{strings.readMore}
+					{DesignHelpers.materialIcon("fullscreen")}
+				</a>
+			</div>
 		</div>
 	);
 };
